pyproject.toml
[project]
name = "local_read_mcp"
version = "0.1.0"
description = "A Model Context Protocol server for document processing with vision support"
readme = "README.md"
authors = [
    { name = "ansatz Vibe", email = "ansatzMe@outlook.com" }
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp<2.14.0",
    "mammoth>=1.9.0",
    "markdownify>=0.14.1",
    "openpyxl>=3.1.5",
    "pdfminer.six",
    "python-pptx",
    "markitdown>=0.1.1",
    "pyyaml>=6.0.2",
    "beautifulsoup4>=4.12.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",  # Vision API (OpenAI-compatible)
    "aiohttp>=3.9.0",  # Vision API
    # MinerU-inspired PDF processing enhancements
    "pypdfium2>=4.30.0",
    "pypdf>=5.6.0",
    "numpy>=1.21.6",
]

[project.scripts]
local-read-mcp = "local_read_mcp.cli:main"
local-read-mcp-server = "local_read_mcp.server.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
# Kept for backward compatibility; PDF deps are now included by default.
pdf = []
# MinerU backend - high-quality PDF parsing with layout analysis, formula recognition, and table recognition
mineru = [
    "mineru>=1.0.0",
]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "coverage[toml]>=7.6.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
minversion = "8.3.5"
testpaths = ["src/test"]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-rA",
    "--show-capture=stderr",
    "--cov=src/local_read_mcp",
    "--cov-report=term-missing",
    "--cov-report=html",
]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.coverage.run]
source = ["src/local_read_mcp"]
omit = [
    "*/test/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
indent-width = 4
